requirements.txt
Flask
pandas
pdfplumber
openpyxl
xlsxwriter
google-generativeai
werkzeug